benchdnn: utils: engine: move to a dedicated file
