+++ tests/benchdnn/dnnl_common.hpp
@@ -33,6 +33,7 @@
 #include "utils/compare.hpp"
 #include "utils/dims.hpp"
 #include "utils/dnnl_query.hpp"
+#include "utils/engine.hpp"
 #include "utils/fill.hpp"
 #include "utils/impl_filter.hpp"
 #include "utils/numeric.hpp"
@@ -41,41 +42,6 @@
 #include "tests/test_thread.hpp"
 
 #define for_ for
-
-#define DNN_SAFE(f, s) \
-    do { \
-        dnnl_status_t status__ = f; \
-        if (status__ != dnnl_success) { \
-            if ((s) == CRIT || (s) == WARN) { \
-                BENCHDNN_PRINT(0, \
-                        "Error: Function '%s' at (%s:%d) returned '%s'\n", \
-                        __FUNCTION__, __FILE__, __LINE__, \
-                        status2str(status__)); \
-                fflush(0); \
-                if ((s) == CRIT) exit(2); \
-            } \
-            return FAIL; \
-        } \
-    } while (0)
-
-#define DNN_SAFE_V(f) \
-    do { \
-        dnnl_status_t status__ = (f); \
-        if (status__ != dnnl_success) { \
-            BENCHDNN_PRINT(0, \
-                    "Error: Function '%s' at (%s:%d) returned '%s'\n", \
-                    __FUNCTION__, __FILE__, __LINE__, status2str(status__)); \
-            fflush(0); \
-            exit(2); \
-        } \
-    } while (0)
-
-// Unlike `DNN_SAFE` this one returns `dnnl_status_t`, not `OK/FAIL`.
-#define DNN_SAFE_STATUS(f) \
-    do { \
-        dnnl_status_t status__ = (f); \
-        if (status__ != dnnl_success) { return status__; } \
-    } while (0)
 
 #ifndef DNNL_EXPERIMENTAL_PROFILING
 #if DNNL_GPU_RUNTIME == DNNL_RUNTIME_OCL \
@@ -91,32 +57,9 @@ extern "C" dnnl_status_t dnnl_query_profiling_data(dnnl_stream_t stream,
 int check_pd_cache(const_dnnl_primitive_desc_t pd, res_t *res);
 int check_primitive_cache(dnnl_primitive_t p, res_t *res);
 
-extern dnnl_engine_kind_t engine_tgt_kind;
-extern size_t engine_index;
 extern isa_hints_t hints;
 extern int default_num_streams;
 extern int num_streams;
-
-struct engine_t {
-    engine_t(dnnl_engine_kind_t engine_kind);
-    engine_t(dnnl_engine_t engine);
-    engine_t(const dnnl::engine &engine);
-    // `recreate_on_copy=true` forces to construct a brand new engine_t object
-    // with underlying interop objects from `other`.
-    // When `recreate_on_copy=false`, copy follows a weak_ptr semantics that
-    // `dnnl::engine` provides.
-    engine_t(const engine_t &other, bool recreate_on_copy = false);
-    operator dnnl_engine_t() const { return engine_.get(); }
-    operator const dnnl::engine &() const { return engine_; }
-
-    bool is_cpu() const;
-    bool is_gpu() const;
-
-private:
-    dnnl::engine::kind get_kind() const;
-    engine_t &operator=(engine_t &other) = delete;
-    dnnl::engine engine_;
-};
 
 struct stream_t {
     stream_t() = default;
